# docs/openapi.yaml
openapi: 3.1.0
info:
  title: icoo_llm_bridge Management API
  version: 2.0.1
  description: |
    Local-first management API used by icoo_desktop. Proxy-compatible LLM
    request and response schemas are intentionally documented separately from
    this management contract.
servers:
  - url: http://127.0.0.1:18181
    description: Default local Bridge
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /api/v1/plugins:
    get:
      operationId: listPlugins
      tags: [Plugins]
      summary: List configured process plugins and runtime status
      responses:
        "200":
          $ref: "#/components/responses/Paged"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/plugins/{id}/ui:
    get:
      operationId: proxyPluginUIRoot
      tags: [Plugins]
      summary: Reverse-proxy plugin loopback admin UI (iframe target)
      parameters:
        - $ref: "#/components/parameters/ResourceID"
      responses:
        "200":
          description: Plugin UI HTML or asset
        "502":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  parameters:
...
    ResourceID:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
...
    Paged:
      description: Paginated response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PagedEnvelope"
    Error:
      description: Request failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Missing or invalid admin API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    PagedEnvelope:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [items, total, page, page_size]
          properties:
            items:
              type: array
              items: {}
            total:
              type: integer
            page:
              type: integer
            page_size:
              type: integer
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
...